Feature/wiki tool (#29)
* Add ChatHeader component

Introduces a new ChatHeader component for the chat interface, providing controls for toggling the sidebar and summary panel, and displaying the active document title.

* Add ChatInput component for chat interface

Introduces the ChatInput component with support for study mode toggle, dynamic input placeholder, and send button. Enhances user interaction in the chat interface by providing contextual input and feedback.

* Add ChatMessageList component for chat UI

Introduces ChatMessageList.jsx to render chat messages, loading states, and document upload prompts. Handles empty states, user and bot message styling, and integrates upload and loading indicators for improved user experience.

* Update button focus ring offset color

Changed the focus:ring-offset color from slate-900 to white in Button component for improved visual consistency.

* Update ChatInput.jsx

* Update Sidebar.jsx

* Add StudyModeToggle component

Introduces a new StudyModeToggle React component for toggling between study mode and standard RAG in the chat interface. Includes UI elements, icons, and animation for status indication.

* Remove Chat page component

Deleted Chat.jsx from src/pages, removing the main chat page implementation. This may be part of a refactor, feature removal, or migration to a new structure.

* Revise README with setup, API, and config details

Expanded the README to clarify project purpose, backend/frontend structure, environment variable configuration, main API endpoints, authentication flow, and troubleshooting tips. The update provides step-by-step instructions for local development, details on data directories, and notes on optional MongoDB and mail integration.

* Update get_answer_from_pdf call with study_mode

Passes study_mode and uses keyword arguments for get_answer_from_pdf in chat_with_pdf endpoint to improve clarity and support additional functionality.

* Add new dependencies for search and tools

Added wikipedia, duckduckgo-search, langchain-tools, and ddgs to requirements.txt to support new features related to search and tool integration.

* Add study_mode field to QuestionSchema

Introduces a new boolean field 'study_mode' with a default value of False to the QuestionSchema model to support study mode functionality.

* Update llm.py
