+++ client/src/components/authUI/Button.jsx
@@ -7,7 +7,6 @@ export default function Button({
   className = "",
   ...props
 }) {
-  // Changed: focus:ring-offset-slate-900 -> focus:ring-offset-white
   const baseStyles =
     "relative inline-flex items-center justify-center px-6 py-3 text-sm font-medium transition-all duration-200 rounded-lg focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-white disabled:opacity-50 disabled:cursor-not-allowed";
