+++ server/schemas/QuestionSchema.py
@@ -4,3 +4,4 @@ from typing import Optional
 class QuestionSchema(BaseModel):
     question: str
     pdf_id: str
+    study_mode: bool = False
